Roster refactoring
