+++ src/commands/resetlead/commands/ResetLeaderToggleVisibility.ts
@@ -1,19 +1,18 @@
 import * as akairo from "discord-akairo";
 import * as discord from "discord.js";
-import { BotgartCommand } from "../../BotgartCommand";
-import * as L from "../../Locale";
-import * as Util from "../../Util";
-import { WvWMap } from "./ResetRoster";
-import * as ResetUtil from "./ResetUtil";
+import { BotgartCommand } from "../../../BotgartCommand";
+import * as L from "../../../Locale";
+import * as Util from "../../../Util";
+import * as ResetUtil from "../ResetUtil";
 
 /**
  Testcases:
 
  */
-export class RemoveResetLeader extends BotgartCommand {
+export class ToggleResetLeaderVisibility extends BotgartCommand {
     constructor() {
-        super("removeresetlead", {
-                aliases: ["removeresetlead", "rmresetlead"],
+        super("toggleresetleadervisibility", {
+                aliases: ["toggleresetleadervisibility"],
                 quoted: true,
                 args: [
                     {
@@ -35,22 +34,18 @@ export class RemoveResetLeader extends BotgartCommand {
         );
     }
 
-    checkArgs(args) {
-        return !args || !args.weekNumber || !args.player ? L.get(this.helptextKey(), [WvWMap.getMaps().map(m => m.name).join(" | ")]) : undefined;
-    }
-
     command(message: discord.Message, responsible: discord.User, guild: discord.Guild, args: any): void {
         if (args.weekNumber < 0) {
             args.weekNumber = ResetUtil.currentWeek()
         }
-        const dbRoster = this.getBotgartClient().getRoster(guild, args.weekNumber, args.year);
+        const dbRoster = this.getBotgartClient().rosterService.getRoster(guild, args.weekNumber, args.year);
         if (dbRoster !== undefined) {
             const [g, mes, roster] = dbRoster;
             const name: string = args.player instanceof discord.GuildMember ? Util.formatUserPing(args.player.id) : args.player;
-            roster.removeLeadByName(WvWMap.getMapByName(args.map), name);
-            this.reply(message, responsible, L.get("ROSTER_LEAD_REMOVED", [name, args.weekNumber, mes.url]));
+            roster.toggleLeaderVisibility(name)
+            this.reply(message, responsible, L.get("ROSTER_LEAD_VISIBILITY_TOGGLED", [name, args.weekNumber, mes.url]));
         }
     }
 }
 
-module.exports = RemoveResetLeader;
+module.exports = ToggleResetLeaderVisibility;
